# docs/architecture/openapi/speakeasy-api.yaml
openapi: 3.0.3
info:
  title: SpeakEasy API
  version: 0.1.0
  description: 'Machine-readable OpenAPI source of truth for SpeakEasy server-backed Product Base behavior, P0 commercial
    subscription readiness, and P0.1 expression automation training. P0.2/P1/P2 roadmap items are intentionally
    excluded from implementation-level paths until they have owning increments.

    '
  license:
    name: Proprietary
servers:
- url: /v1
  description: Versioned API base path
security:
- bearerAuth: []
...
paths:
...
  /practice/sessions:
    post:
      tags:
      - Practice
      operationId: startPracticeSession
      summary: Start or resume Product Base voice practice session.
      x-traceability:
        product_objects:
        - product-base
        requirements:
        - FR-007
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartSessionRequest'
            examples:
              default:
                $ref: '#/components/examples/StartSessionRequestExample'
      responses:
        '200':
          description: Practice session started or resumed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PracticeSessionResponse'
              examples:
                default:
                  $ref: '#/components/examples/PracticeSessionResponseExample'
        '422':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  $ref: '#/components/examples/ErrorResponseExample'
    parameters:
    - $ref: '#/components/parameters/RequestId'
  /practice/sessions/{session_id}:
    get:
      tags:
      - Practice
      operationId: getPracticeSession
      summary: Get Product Base practice session.
      x-traceability:
        product_objects:
        - product-base
        requirements:
        - FR-007
      parameters:
      - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Practice session state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PracticeSessionResponse'
              examples:
                default:
                  $ref: '#/components/examples/PracticeSessionResponseExample'
        '404':
          $ref: '#/components/responses/NotFound'
    parameters:
    - $ref: '#/components/parameters/RequestId'
  /practice/sessions/{session_id}/turns:
    post:
      tags:
      - Practice
      operationId: submitPracticeTurn
      summary: Submit a Product Base practice turn.
      x-idempotency-required: true
      x-traceability:
        product_objects:
        - product-base
        requirements:
        - FR-007
        - FR-008
        - FR-009
      parameters:
      - $ref: '#/components/parameters/SessionId'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitTurnRequest'
            examples:
              default:
                $ref: '#/components/examples/SubmitTurnRequestExample'
      responses:
        '200':
          description: Practice turn accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PracticeTurnResponse'
              examples:
                default:
                  $ref: '#/components/examples/PracticeTurnResponseExample'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
    parameters:
    - $ref: '#/components/parameters/RequestId'
  /practice/sessions/{session_id}/complete:
    post:
      tags:
      - Practice
      operationId: completePracticeSession
      summary: Complete Product Base practice session and return summary.
      x-traceability:
        product_objects:
        - product-base
        requirements:
        - FR-009
      parameters:
      - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Practice session completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionSummaryResponse'
              examples:
                default:
                  $ref: '#/components/examples/SessionSummaryResponseExample'
        '404':
          description: Referenced resource not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  $ref: '#/components/examples/ErrorResponseExample'
    parameters:
    - $ref: '#/components/parameters/RequestId'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    opsBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Restricted operations token.
    webhookSignature:
      type: apiKey
      in: header
      name: X-Provider-Signature
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
...
    SessionId:
      name: session_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    IdempotencyConflict:
      description: Idempotency key reused with a different body.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              $ref: '#/components/examples/ErrorResponseExample'
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              $ref: '#/components/examples/ErrorResponseExample'
...
  schemas:
    SchemaVersion:
      type: integer
      enum:
      - 1
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - request_id
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
              additionalProperties: true
    ErrorCode:
      type: string
      enum:
      - UNAUTHENTICATED
      - FORBIDDEN
      - ENTITLEMENT_REQUIRED
      - USAGE_LIMIT_EXCEEDED
      - INVALID_RECEIPT
      - PRODUCT_MISMATCH
      - SUBSCRIPTION_EXPIRED
      - IDEMPOTENCY_CONFLICT
      - SCHEMA_VALIDATION_FAILED
      - PROVIDER_UNAVAILABLE
      - DELETE_IN_PROGRESS
      - RESOURCE_NOT_FOUND
      - CONFLICT
    LevelCode:
      type: string
      enum:
      - L1
      - L2
      - L3
    ScenarioId:
      type: string
      enum:
      - job_interview
      - onboarding_introduction
...
    StartSessionRequest:
      type: object
      required:
      - schema_version
      - scenario_id
      - level_code
      properties:
        schema_version:
          $ref: '#/components/schemas/SchemaVersion'
        scenario_id:
          $ref: '#/components/schemas/ScenarioId'
        level_code:
          $ref: '#/components/schemas/LevelCode'
        resume_existing:
          type: boolean
          default: true
    PracticeSessionResponse:
      type: object
      required:
      - schema_version
      - session
      properties:
        schema_version:
          $ref: '#/components/schemas/SchemaVersion'
        session:
          $ref: '#/components/schemas/PracticeSession'
    PracticeSession:
      type: object
      required:
      - session_id
      - scenario_id
      - level_code
      - status
      properties:
        session_id:
          type: string
        scenario_id:
          $ref: '#/components/schemas/ScenarioId'
        level_code:
          $ref: '#/components/schemas/LevelCode'
        status:
          type: string
          enum:
          - active
          - idle
          - recording
          - submitting
          - awaiting_feedback
          - feedback
          - recoverable_error
          - completed
        current_turn_index:
          type: integer
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ConversationMessage'
    SubmitTurnRequest:
      type: object
      required:
      - schema_version
      properties:
        schema_version:
          $ref: '#/components/schemas/SchemaVersion'
        transcript:
          type: string
        audio_ref:
          type: string
        client_state_version:
          type: integer
    PracticeTurnResponse:
      type: object
      required:
      - schema_version
      - session_id
      properties:
        schema_version:
          $ref: '#/components/schemas/SchemaVersion'
        session_id:
          type: string
        session_status:
          type: string
          enum:
          - active
          - feedback
          - recoverable_error
          - completed
        user_message:
          $ref: '#/components/schemas/ConversationMessage'
        coach_feedback:
          $ref: '#/components/schemas/CoachFeedback'
        learning_evidence_candidates:
          type: array
          items:
            $ref: '#/components/schemas/LearningEvidenceCandidate'
        recoverable_error:
          $ref: '#/components/schemas/RecoverableProviderError'
    ConversationMessage:
      type: object
      required:
      - message_id
      - role
      - text
      - created_at
      properties:
        message_id:
          type: string
        role:
          type: string
          enum:
          - learner
          - coach
          - system
          - scenario_partner
        text:
          type: string
        audio_ref:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    CoachFeedback:
      type: object
      required:
      - summary
      properties:
        summary:
          type: string
        feedback_type:
          type: string
          enum:
          - next_question
          - retry
          - expression_suggestion
          - score_signal
          - recoverable_error
        main_issue_type:
          type: string
        suggested_expression:
          type: string
        next_prompt:
          type: string
        score_signal:
          $ref: '#/components/schemas/ScoreSignal'
        validation_status:
          type: string
          enum:
          - valid
          - invalid_schema
          - fallback
        provider_status:
          type: string
          enum:
          - success
          - timeout
          - provider_unavailable
          - invalid_schema
    RecoverableProviderError:
      type: object
      required:
      - code
      - message
      - retryable
      properties:
        code:
          type: string
        message:
          type: string
        retryable:
          type: boolean
    ScoreSignal:
      type: object
      required:
      - score_kind
      - status
      - source
      properties:
        score_kind:
          type: string
          enum:
          - pronunciation
          - completeness
          - grammar
          - task_completion
          - fluency
        value:
          type: number
          minimum: 0
          maximum: 1
          nullable: true
        confidence:
          type: number
          minimum: 0
          maximum: 1
          nullable: true
        status:
          type: string
          enum:
          - available
          - low_confidence
          - unavailable
        source:
          type: string
          enum:
          - server_side_adapter
    SessionSummaryResponse:
      type: object
      required:
      - schema_version
      - summary
      properties:
        schema_version:
          $ref: '#/components/schemas/SchemaVersion'
        summary:
          $ref: '#/components/schemas/SessionSummary'
    SessionSummary:
      type: object
      required:
      - session_id
      - learned_items
      - weak_points
      - next_focus
      properties:
        session_id:
          type: string
        learned_items:
          type: array
          items:
            type: string
        weak_points:
          type: array
          items:
            type: string
        next_focus:
          type: string
...
    LearningEvidenceCandidate:
      type: object
      required:
      - candidate_id
      - evidence_type
      - confidence
      - status
      properties:
        candidate_id:
          type: string
        evidence_type:
          type: string
          enum:
          - mastered_expression
          - weak_expression
          - correction
          - saved_expression
          - review_due
        target_expression_id:
          type: string
        text:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        status:
          type: string
          enum:
          - candidate
          - accepted
          - rejected
          - merged_duplicate
...
  examples:
    ErrorResponseExample:
      summary: Standard error response
      value:
        error:
          code: SCHEMA_VALIDATION_FAILED
          message: Request validation failed.
          request_id: req_20260526_sample
          details:
            field: schema_version
            reason: Unsupported schema version.
...
    StartSessionRequestExample:
      summary: startPracticeSession request example
      value:
        schema_version: 1
        scenario_id: job_interview
        level_code: L1
        resume_existing: true
    PracticeSessionResponseExample:
      summary: startPracticeSession response 200 example
      value:
        schema_version: 1
        session:
          session_id: session_id_sample
          scenario_id: job_interview
          level_code: L1
          status: active
          current_turn_index: 1
          messages:
          - message_id: message_id_sample
            role: learner
            text: Sample text.
            audio_ref: sample
            created_at: '2026-05-26T12:00:00Z'
    SubmitTurnRequestExample:
      summary: submitPracticeTurn request example
      value:
        schema_version: 1
        transcript: Sample text.
        audio_ref: sample
        client_state_version: 1
    PracticeTurnResponseExample:
      summary: submitPracticeTurn response 200 example
      value:
        schema_version: 1
        session_id: session_id_sample
        session_status: feedback
        user_message:
          message_id: message_id_sample
          role: learner
          text: Sample text.
          audio_ref: sample
          created_at: '2026-05-26T12:00:00Z'
        coach_feedback:
          summary: sample
          feedback_type: next_question
          main_issue_type: sample
          suggested_expression: sample
          next_prompt: sample
          validation_status: valid
          provider_status: success
          score_signal:
            score_kind: pronunciation
            value: 0.85
            confidence: 0.85
            status: available
            source: server_side_adapter
        learning_evidence_candidates:
        - candidate_id: candidate_id_sample
          evidence_type: mastered_expression
          target_expression_id: target_expression_id_sample
          text: Sample text.
          confidence: 0.85
          status: candidate
    SessionSummaryResponseExample:
      summary: completePracticeSession response 200 example
      value:
        schema_version: 1
        summary:
          session_id: session_id_sample
          learned_items:
          - sample
          weak_points:
          - sample
          next_focus: sample
...